refactor UserPreviewDTO - refactor star type to double

+++ src/main/java/loty/lostem/dto/UserPreviewDTO.java
@@ -13,7 +13,7 @@ public class UserPreviewDTO {
 
     private String profile;
 
-    private float star;
+    private double star;
 
     private String tag;
 }
